+++ packages/opa/src/funcs/executeBatchPolicyWithInput.ts
@@ -125,7 +125,7 @@ async function $do(
   const context = {
     baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "executeBatchPolicyWithInput",
-    oAuth2Scopes: null,
+    oAuth2Scopes: [],
 
     resolvedSecurity: requestSecurity,