+++ packages/opa/src/lib/config.ts
@@ -57,7 +57,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.2.0",
-  sdkVersion: "1.7.4",
-  genVersion: "2.562.2",
-  userAgent: "speakeasy-sdk/typescript 1.7.4 2.562.2 0.2.0 @styra/opa",
+  sdkVersion: "1.7.5",
+  genVersion: "2.562.3",
+  userAgent: "speakeasy-sdk/typescript 1.7.5 2.562.3 0.2.0 @styra/opa",
 } as const;